requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11.4
django-filter==1.0.4
djangorestframework==3.6.3
future==0.16.0
gunicorn==19.7.1
idna==2.5
line-bot-sdk==1.3.0
Markdown==2.6.8
pytz==2017.2
requests==2.18.3
sqlite3==0.0.0
urllib3==1.22
whitenoise==3.3.0
